Add support for more than one track
VLC doesn't understand streams very well because there is no end time
and no "Ended" state. Instead of waiting for the song to end on its
own, we grab the song duration from the track API and use that to set a
timeout on the player. Then we can iterate over each song in the track
list like a makeshift playlist.