Fix angry vlc errors
VLC gets upset when you call player.play() repeatedly in a loop and if
these errors aren't handled it can fill up cron mail. Instead of
calling play over and over, call it once and then put a sleep in a loop
so that the program blocks while the song is playing.